chore: move RgbPixels to it's own module
